Added hooks for cleanlinesss
