+++ src/modules/modules.tsx
@@ -1,4 +1,4 @@
-export interface Users {
+export interface Players {
 	player1: string;
 	player2: string;
 }
